Make introspection compile with -Wall -Werror
Address all gcc warnings, -Werror is not enabled yet but at least
-Wall -Werror passes on my machine.

+++ tests/gimarshallingtests.c
@@ -3826,7 +3826,7 @@ gi_marshalling_tests_object_vfunc_with_callback (GIMarshallingTestsObject *objec
 static int
 _callback (int val, void *user_data)
 {
-    g_assert(user_data == 0xdeadbeef);
+    g_assert((int)user_data == (int)0xdeadbeef);
     return val;
 }
